+++ RELEASENOTES.md
@@ -10,11 +10,10 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 * Feature - Add download document in app detail page
 * Feature - App Overview Page - Design updates
 * Feature - Assign User Role
+* Feature - App overview - Show overlay when Application status is in "in progress"
 * Feature - Connect My App API on Home Page
 * Feature - App Release Process - Business logic & API connect for document upload
 * Feature - My App User Experience
-
-
 
 ### 0.5.5